+++ components/layout/TableSkeleton.tsx
@@ -1,4 +1,4 @@
-export function TableSkeleton() {
+export default function TableSkeleton() {
   return (
     <div className="animate-pulse">
       <div className="h-8 bg-gray-300 mb-4"></div>
